pyproject.toml
[tool.poetry]
name = "iot-controller"
version = "0.1.0"
description = ""
authors = ["Timothy Holmes <tim.a.holmes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.81.0"
PyP100 = "^0.0.19"
python-box = {extras = ["all"], version = "^6.0.2"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.8.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
